requirement.txt
alembic==1.5.4
blinker==1.4
captcha==0.3
click==7.1.2
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
-e git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc93e684fa0a3a4350a38e41ae00483a796#egg=Flask_Uploads
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
Pillow==8.1.0
psutil==5.8.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
Werkzeug==1.0.1
